restructured sideNav
