+++ src/components/SideNav/SearchBar.js
