+++ src/components/SideNav/SideNav.js
@@ -1,11 +1,28 @@
 import React from "react"
-import SideNavCategories from './SideNavCategories'
-import {ButtonGroup} from 'react-bootstrap';
+import {ButtonGroup, Button, DropdownButton} from 'react-bootstrap';
+import SearchBar from './SearchBar.js'
 
 const SideNav = () => {
         return (
             <ButtonGroup vertical>
-                <SideNavCategories/>
+              <Button>See All Cheeses</Button>
+              <DropdownButton title="Search By Firmness" id="bg-vertical-dropdown-1">
+                <SearchBar/>
+              </DropdownButton>
+              <DropdownButton title="Search By Animal" id="bg-vertical-dropdown-1">
+                <SearchBar/>
+              </DropdownButton>
+              <DropdownButton title="Search By Cheese Name" id="bg-vertical-dropdown-1">
+                <SearchBar/>
+              </DropdownButton>
+              <DropdownButton title="Search By Substitute" id="bg-vertical-dropdown-1">
+                <SearchBar/>
+              </DropdownButton>
+              <Button>Find in Stores</Button>
+              <DropdownButton title="Give Me That Random Cheese!" id="bg-vertical-dropdown-1">
+                <SearchBar/>
+              </DropdownButton>
+              <Button>See My Favorites</Button>
             </ButtonGroup>
         )
