+++ src/components/action/index.js
@@ -16,5 +16,5 @@ function fetchCheese(cheeseName) {
  fetch(`http://cheeswhiz.herokuapp.com/api/cheese/firmness/${cheeseName}`)
  .then(response => response.json())
  .then(result => console.log(result))
- .catch((err) => console.log('Houston we has a prolem!'))
+ .catch((err) => console.log('Houston we has a problem!'))
 }
